Software/Website/flaskapp/requirements.txt
Flask~=3.0
flask-cors~=4.0
python-dotenv~=1.0
bcrypt~=4.1
pytest~=8.0
